rename folder place to sight and repository to sights
