+++ lib/data/repository/sight_repository.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
 import 'package:dio/dio.dart';
-import 'package:places/data/model/places/sight_dto.dart';
-import 'package:places/data/model/places/sights_filter_request_dto.dart';
+import 'package:places/data/model/sights/sight_dto.dart';
+import 'package:places/data/model/sights/sights_filter_request_dto.dart';
 import 'package:retrofit/retrofit.dart';
 
 import '../model/sight.dart';
